Refactor: move result item mapper to core mappers and remove unused dependency
