+++ core/movie/build.gradle.kts
@@ -8,10 +8,9 @@ android.namespace = "com.waffiq.bazz_movies.core.movie"
 
 dependencies {
   api(project(":core:common"))
-  api(project(":core:database"))
   api(project(":core:domain"))
+  api(project(":core:mappers"))
   api(project(":core:network"))
-  implementation(project(":core:mappers"))
 
   implementation(libs.androidx.core.ktx)
   implementation(libs.androidx.lifecycle.common)
