+++ core/movie/src/main/kotlin/com/waffiq/bazz_movies/core/movie/utils/mappers/Mapper.kt
@@ -1,9 +1,6 @@
 package com.waffiq.bazz_movies.core.movie.utils.mappers
 
-import com.waffiq.bazz_movies.core.common.utils.Constants.MOVIE_MEDIA_TYPE
-import com.waffiq.bazz_movies.core.domain.ResultItem
 import com.waffiq.bazz_movies.core.movie.domain.model.post.PostFavoriteWatchlist
-import com.waffiq.bazz_movies.core.network.data.remote.responses.tmdb.ResultItemResponse
 import com.waffiq.bazz_movies.core.network.data.remote.responses.tmdb.post.PostFavoriteWatchlistResponse
 
 object Mapper {
@@ -11,26 +8,5 @@ object Mapper {
   fun PostFavoriteWatchlistResponse.toPostFavoriteWatchlist() = PostFavoriteWatchlist(
     statusCode = statusCode,
     statusMessage = statusMessage
-  )
-
-  fun ResultItemResponse.toResultItem(): ResultItem = ResultItem(
-    firstAirDate = firstAirDate,
-    overview = overview,
-    originalLanguage = originalLanguage,
-    listGenreIds = genreIds,
-    posterPath = posterPath,
-    backdropPath = backdropPath,
-    mediaType = mediaType ?: MOVIE_MEDIA_TYPE,
-    originalName = originalName,
-    popularity = popularity,
-    voteAverage = voteAverage,
-    name = name,
-    id = id ?: 0,
-    adult = adult == true,
-    voteCount = voteCount ?: 0,
-    originalTitle = originalTitle,
-    video = video == true,
-    title = title,
-    releaseDate = releaseDate
   )
 }
