+++ feature/detail/build.gradle.kts
@@ -10,7 +10,7 @@ plugins {
 android.namespace = "com.waffiq.bazz_movies.feature.detail"
 
 dependencies {
-  implementation(project(":core:domain"))
+  implementation(project(":core:database"))
   implementation(project(":core:movie"))
   implementation(project(":core:user"))
   implementation(project(":core:uihelper"))
