+++ feature/detail/src/main/kotlin/com/waffiq/bazz_movies/feature/detail/data/repository/DetailRepositoryImpl.kt
@@ -4,7 +4,7 @@ import androidx.paging.PagingData
 import androidx.paging.map
 import com.waffiq.bazz_movies.core.domain.Outcome
 import com.waffiq.bazz_movies.core.domain.ResultItem
-import com.waffiq.bazz_movies.core.movie.utils.mappers.Mapper.toResultItem
+import com.waffiq.bazz_movies.core.mappers.ResultItemMapper.toResultItem
 import com.waffiq.bazz_movies.core.network.data.remote.datasource.MovieDataSource
 import com.waffiq.bazz_movies.core.network.utils.result.NetworkResult
 import com.waffiq.bazz_movies.feature.detail.domain.model.MovieTvCredits
