+++ feature/login/build.gradle.kts
@@ -9,9 +9,10 @@ plugins {
 android.namespace = "com.waffiq.bazz_movies.feature.login"
 
 dependencies {
+  implementation(project(":core:common"))
+  implementation(project(":core:domain"))
   implementation(project(":core:uihelper"))
   implementation(project(":core:user"))
-  implementation(project(":core:movie"))
 
   implementation(libs.androidx.appcompat)
   implementation(libs.google.material)
